use uv for project management
use static versions of sox and ffmpeg

fix static binary loading

more changes

update shell script

update colab notebook

suppress uv logging in notebook

remove unused files and rename urvc.sh to urvc

update readme

small updates to todo list
