+++ README.md
@@ -14,14 +14,10 @@ Ultimate RVC is under constant development and testing, but you can try it out r
 * Caching system which saves intermediate audio files as needed, thereby reducing inference time as much as possible. For example, if song A has already been converted using model B and now you want to convert song A using model C, then vocal extraction can be skipped and inference time reduced drastically
 * Ability to listen to intermediate audio files in the UI. This is useful for getting an idea of what is happening in each step of the song cover generation pipeline
 * A "multi-step" song cover generation tab: here you can try out each step of the song cover generation pipeline in isolation. For example, if you already have extracted vocals available and only want to convert these using your voice model, then you can do that here. Besides, this tab is useful for experimenting with settings for each step of the song cover generation pipeline
-* An overhaul of the song input component for the song cover generation pipeline. Now cached input songs can be selected from a dropdown, so that you don't have to supply the Youtube link of a song each time you want to convert it. 
+* An overhaul of the song input component for the song cover generation pipeline. Now cached input songs can be selected from a dropdown, so that you don't have to supply the Youtube link of a song each time you want to convert it.
 * A new "manage models" tab, which collects and revamps all existing functionality for managing voice models, as well as adds some new features, such as the ability to delete existing models
-* A "manage audio", which allows you to to interact with all audio generated by the app. Currently, this tab supports deleting audio files.
-* Lots of visual and performance improvements resulting from updating from Gradio 3 to Gradio 4 and from python 3.9 to python 3.11
-
-## PRO TIP: Use a GPU for faster processing
-
-While it is possible to run the Ultimate RVC web app on a CPU, it is highly recommended to use a GPU for faster processing. On an NVIDIA 3080 GPU, the AI cover generation process takes approximately 1.5 minutes, while on a CPU, it takes approximately 15 minutes. No testing has been done on AMD GPUs, so no guarantees are made for their performance.
+* A new "manage audio" tab, which allows you to interact with all audio generated by the app. Currently, this tab supports deleting audio files.
+* Lots of visual and performance improvements resulting from updating from Gradio 3 to Gradio 4 and from python 3.9 to python 3.12
 
 ## Colab notebook
 
@@ -29,17 +25,28 @@ For those without a powerful enough NVIDIA GPU, you may try Ultimate RVC out usi
 
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/JackismyShephard/ultimate-rvc/blob/main/notebooks/ultimate_rvc_colab.ipynb)
 
-For those who want to run this locally, follow the setup guide below.
+For those who want to run the Ultimate RVC project locally, follow the setup guide below.
 
 ## Setup
+
+The Ultimate RVC project currently supports Windows and Debian-based Linux distributions, namely Ubuntu 22.04 and Ubuntu 24.04. Support for other platforms is not guaranteed.
+
+To setup the project follow the steps below and execute the provided commands in an appropriate terminal. On windows this terminal should be **powershell**, while on debian-based linux distributions it should be a **bash**-compliant shell.
 
 ### Install Git
 
 Follow the instructions [here](https://git-scm.com/book/en/v2/Getting-Started-Installing-Git) to install Git on your computer.
 
-### Clone Ultimate RVC repository
+### Set execution policy (Windows only)
 
-Open a terminal and run the following commands to clone this entire repository and open it locally.
+To execute the subsequent commands on Windows, it is necessary to first grant
+powershell permission to run scripts. This can be done on a user-level as follows:
+
+```console
+Set-ExecutionPolicy RemoteSigned -Scope CurrentUser
+```
+
+### Clone Ultimate RVC repository
 
 ```console
 git clone https://github.com/JackismyShephard/ultimate-rvc
@@ -48,42 +55,17 @@ cd ultimate-rvc
 
 ### Install dependencies
 
-#### Windows
-
-Run the following command in powershell to install the necessary dependencies on Windows:
-
-```bat
-./urvc.bat install 
+```console
+./urvc install 
 ```
-
-Note that this will install Miniconda in your user directory. 
-The whole process may take upwards of 10 minutes, so grab a cup of coffee and wait.
-
-#### Linux (Debian-based)
-
-Run the following command to install the necessary dependencies on Debian-based Linux distributions (e.g. Ubuntu):
-
-```bash
-./urvc.sh install 
-```
-
-The command has been tested only on Ubuntu 22.04 and 24.04 so support for other distributions is not guaranteed. 
-Also note that the command will install the CUDA 12.1 toolkit system-wide. In case you have problems, you may need to install the toolkit manually.
+Note that on Linux, this command will install the CUDA 12.4 toolkit system-wide, if it is not already available. In case you have problems, you may need to install the toolkit manually.
 
 ## Usage
 
 ### Start the app
 
-#### Windows
-
-```bat
-./urvc.bat run
-```
-
-#### Linux (Debian-based)
-
-```bash
-./urvc.sh run
+```console
+./urvc run
 ```
 
 Once the following output message `Running on local URL:  http://127.0.0.1:7860` appears, you can click on the link to open a tab with the web app.
@@ -102,7 +84,7 @@ Once the 2 input fields are filled in, simply click `Download`! Once the output 
 
 #### Upload models
 
-![](images/webui_upload_model.png?raw=true)
+![ ](images/webui_upload_model.png?raw=true)
 
 For people who have trained RVC v2 models locally and would like to use them for AI cover generations.
 Navigate to the `Upload model` subtab under the `Manage models` tab, and follow the instructions.
@@ -157,16 +139,8 @@ The directory structure should look something like this:
 
 #### Usage
 
-##### Windows
-
-```bat
-./urvc.bat cli song-cover run-pipeline [OPTIONS] SOURCE MODEL_NAME 
-```
-
-##### Linux (Debian-based)
-
-```bash
-./urvc.sh cli song-cover run-pipeline [OPTIONS] SOURCE MODEL_NAME
+```console
+./urvc cli song-cover run-pipeline [OPTIONS] SOURCE MODEL_NAME 
 ```
 
 ##### Arguments
@@ -198,35 +172,16 @@ The directory structure should look something like this:
 
 ## Update to latest version
 
-Run the following command to pull latest changes from the repository and reinstall dependencies.
-Note that the process may take upwards of 5 minutes.
-
-### Windows
-
-```bat
-./urvc.bat update
-```
-
-### Linux (Debian-based)
-
-```bash
-./urvc.sh update
+```console
+./urvc update
 ```
 
 ## Development mode
 
 When developing new features or debugging, it is recommended to run the app in development mode. This enables hot reloading, which means that the app will automatically reload when changes are made to the code.
 
-### Windows
-
-```bat
-./urvc.bat dev
-```
-
-### Linux (Debian-based)
-
-```bash
-./urvc.sh dev
+```console
+./urvc dev
 ```
 
 ## Terms of Use
