+++ pyproject.toml
@@ -1,9 +1,106 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "ultimate-rvc"
+version = "0.1.1"
+description = "Ultimate RVC"
+readme = "README.md"
+requires-python = "==3.12.*"
+dependencies = [
+    # General
+    "lib==4.0.0",
+    "pydantic==2.9.2",
+    "typer==0.12.5",
+    
+    # Networking
+    "requests==2.32.3",
+    "yt_dlp==2024.8.6",
+    # TODO add these later
+    # "deemix",
+    # "wget",
+    # "flask",
+    # "beautifulsoup4",
+    # "pypresence",
+    # Data science
+    "numpy==1.26.4",
+    "scipy==1.14.1",
+    "matplotlib==3.9.2",
+    "tqdm==4.66.5",
+    "gradio==4.44.1",
+    
+    # Machine learning
+    "torch==2.4.1+cu124",
+    "torchaudio==2.4.1+cu124",
+    "torchcrepe==0.0.23",
+    "fairseq",
+    "faiss-cpu==1.9.0",
+    "tensorboardX==2.6.2.2",
+    # TODO add these later
+    # "tensorboard",
+    # "torchfcpe",
+    # "local-attention",
+    # "libf0",
+    # "einops",
+    # "numba; sys_platform == 'linux'",
+    # "numba==0.57.0; sys_platform == 'darwin' or sys_platform == 'win32'",
+    
+    # Audio
+    "static-ffmpeg==2.7",
+    "static-sox==1.0.1",
+    "typed-ffmpeg==2.0.0",
+    "soundfile==0.12.1",
+    "librosa==0.10.2",
+    "sox==1.5.0",
+    "pydub==0.25.1",
+    "pydub-stubs==0.25.1.1",
+    "pedalboard==0.9.16",
+    "audio-separator[gpu]==0.21.2",
+    "praat-parselmouth==0.4.5",
+    "pyworld==0.3.4",
+    "diffq==0.2.4",
+    # TODO add the later
+    # "noisereduce",
+    # "audio_upscaler==0.1.4",
+    # "edge-tts==6.1.9",
+    # "ffmpeg-python>=0.2.0",
+    # "ffmpy==0.3.1"
+]
+
+[project.scripts]
+urvc = "cli.main:app"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/cli"]
+
+[tool.uv]
+environments = ["sys_platform == 'win32'", "sys_platform == 'linux'"]
+cache-dir = "./dependencies/uv/cache"
+compile-bytecode = true
+
+[tool.uv.sources]
+torch = { index = "torch-cu124"}
+torchaudio = { index = "torch-cu124"}
+fairseq = [
+    { path = "./dependencies/wheels/fairseq-0.12.3.1-cp312-cp312-win_amd64.whl", marker = "sys_platform == 'win32'" },
+    { path = "./dependencies/wheels/fairseq-0.12.3.1-cp312-cp312-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
+]
+diffq = { path = "./dependencies/wheels/diffq-0.2.4-cp312-cp312-win_amd64.whl", marker = "sys_platform == 'win32'"}
+pyworld = { path = "./dependencies/wheels/pyworld-0.3.4-cp312-cp312-win_amd64.whl", marker = "sys_platform == 'win32'"}
+
+[[tool.uv.index]]
+name = "torch-cu124"
+url = "https://download.pytorch.org/whl/cu124"
+explicit = true
+
 [tool.pyright]
 stubPath = "src/stubs"
 pythonVersion = "3.12"
 pythonPlatform = "All"
 typeCheckingMode = "strict"
 ignore = ["**/.venv"]
+exclude = ["./dependencies"]
 
 [tool.black]
 target-version = ['py312']
@@ -67,9 +164,9 @@ section-order = [
     "first-party",
     "base",
     "vc",
-    "backend",
-    "frontend",
-    "main",
+    "core",
+    "web",
+    "cli",
     "local-folder",
 ]
 
@@ -122,9 +219,9 @@ section-order = [
 ]
 "base" = ["common", "typing_extra"]
 "vc" = ["vc"]
-"backend" = ["backend"]
-"frontend" = ["frontend"]
-"main" = ["app", "cli", "init"]
+"core" = ["core"]
+"web" = ["web"]
+"cli" = ["cli"]
 
 [tool.ruff.lint.pycodestyle]
 max-doc-length = 72
