--- requirements.txt
@@ -1,61 +1,0 @@
-# General
-lib==4.0.0
-pydantic==2.9.2
-typer==0.12.5
-
-# Networking
-requests==2.32.3 #NOTE upgraded from 2.32.0
-yt_dlp==2024.8.6
-#TODO add these later
-# deemix
-# wget
-# flask
-# beautifulsoup4
-# pypresence
-
-# Data science
-numpy==1.26.4
-scipy==1.14.1 # NOTE upgraded from 1.11.1
-matplotlib==3.9.2 #NOTE upgraded from 3.7.2
-tqdm==4.66.5 #NOTE upgraded from unspecified
-gradio==4.44.1
-
-# Machine learning
---find-links https://download.pytorch.org/whl/torch/
-torch==2.4.1+cu124
---find-links https://download.pytorch.org/whl/torchaudio/
-torchaudio==2.4.1+cu124
-torchcrepe==0.0.23 # NOTE upgraded from 0.0.20
-./dependencies/wheels/fairseq-0.12.3.1-cp312-cp312-win_amd64.whl; sys_platform == 'win32'
-./dependencies/wheels/fairseq-0.12.3.1-cp312-cp312-linux_x86_64.whl; sys_platform == 'linux'
-./dependencies/wheels/diffq-0.2.4-cp312-cp312-win_amd64.whl; sys_platform == 'win32'
-.; sys_platform == 'linux'
-faiss-cpu==1.9.0
-tensorboardX==2.6.2.2
-#TODO add these later
-# tensorboard
-# torchfcpe
-# local-attention
-# libf0 
-# einops 
-# numba; sys_platform == 'linux'
-# numba==0.57.0; sys_platform == 'darwin' or sys_platform == 'win32'
-
-# Audio
-typed-ffmpeg==2.0.0
-soundfile==0.12.1
-librosa==0.10.2 # NOTE upgraded from 0.9.2
-sox==1.5.0
-pydub==0.25.1
-pydub-stubs==0.25.1.1
-pedalboard==0.9.16
-audio-separator[gpu]==0.21.2
-praat-parselmouth==0.4.5 # NOTE upgraded from unspecified
-pyworld==0.3.4; sys_platform == 'linux'
-./dependencies/wheels/pyworld-0.3.4-cp312-cp312-win_amd64.whl; sys_platform == 'win32'
-#TODO add the later
-# noisereduce
-# audio_upscaler==0.1.4
-# edge-tts==6.1.9
-# ffmpeg-python>=0.2.0
-# ffmpy==0.3.1
