--- src/backend/__init__.py
@@ -1,7 +1,0 @@
-"""
-Backend package for the Ultimate RVC project.
-
-This package contains modules for managing audio files, voice models,
-and song cover generation
-
-"""
