+++ src/cli/__init__.py
@@ -1,0 +1,4 @@
+"""
+Package which defines the command-line interface for the Ultimate RVC
+project.
+"""
