+++ src/core.py
@@ -1,0 +1,9 @@
+"""
+Main entry point for the initialization of the Ultimate RVC
+project.
+"""
+
+from core.main import initialize
+
+if __name__ == "__main__":
+    initialize()
