+++ src/core/__init__.py
@@ -1,0 +1,13 @@
+"""
+core package for the Ultimate RVC project.
+
+This package contains modules for managing audio files, voice models,
+and song cover generation.
+
+"""
+
+import static_ffmpeg
+import static_sox
+
+static_ffmpeg.add_paths()
+static_sox.add_paths()
