+++ src/stubs/static_ffmpeg/__init__.pyi
@@ -1,0 +1,1 @@
+def add_paths(weak: bool = False) -> bool: ...
