+++ src/stubs/static_sox/__init__.pyi
@@ -1,0 +1,1 @@
+def add_paths(weak: bool = False) -> bool: ...
