+++ src/web/tabs/__init__.py
@@ -1,5 +1,5 @@
 """
-Frontend package for tabs.
+web package for tabs.
 
 This package contains modules which define functions for rendering each
 tab in the gradio-based UI of the Ultimate RVC project.
